+++ alacritty_terminal/src/term/mod.rs
@@ -1169,7 +1169,7 @@ impl<T: EventListener> Handler for Term<T> {
             Some('>') => {
                 trace!("Reporting secondary device attributes");
                 let version = version_number(env!("CARGO_PKG_VERSION"));
-                let text = format!("\x1b[>0;{};1c", version);
+                let text = format!("\x1b[>0;{version};1c");
                 self.event_proxy.send_event(Event::PtyWrite(text));
             },
             _ => debug!("Unsupported device attributes intermediate"),
@@ -1944,7 +1944,7 @@ impl<T: EventListener> Handler for Term<T> {
         self.event_proxy.send_event(Event::TextAreaSizeRequest(Arc::new(move |window_size| {
             let height = window_size.num_lines * window_size.cell_height;
             let width = window_size.num_cols * window_size.cell_width;
-            format!("\x1b[4;{};{}t", height, width)
+            format!("\x1b[4;{height};{width}t")
         })));
     }
 
@@ -2721,7 +2721,7 @@ mod tests {
         term.reset_damage();
         let vi_cursor_point = term.vi_mode_cursor.point;
         let line = vi_cursor_point.line.0 as usize;
-        let left = vi_cursor_point.column.0 as usize;
+        let left = vi_cursor_point.column.0;
         let right = left;
 
         let mut damaged_lines = match term.damage(None) {