+++ alacritty_terminal/tests/ref.rs
@@ -126,13 +126,7 @@ fn ref_test(dir: &Path) {
                 let cell = &term_grid[Line(i as i32)][Column(j)];
                 let original_cell = &grid[Line(i as i32)][Column(j)];
                 if original_cell != cell {
-                    println!(
-                        "[{i}][{j}] {original:?} => {now:?}",
-                        i = i,
-                        j = j,
-                        original = original_cell,
-                        now = cell,
-                    );
+                    println!("[{i}][{j}] {original_cell:?} => {cell:?}",);
                 }
             }
         }